Add support for functions
Inspired by SubX functions: http://akkartik.name/post/mu-2019-1

Format:
(<reg> <function> [<arg>[, <arg>[, ...]])

Key differences with SubX:
* <reg> arg is required since SubX doesn't have dedicated sp.
  Anything that can be the destination of a copy instruction is
  supported.
* <args> are comma separated, anything that can be the source
  of a copy instruction is allowed.

Currently, the following instructions are generated by this command:

0/push address/ 0.reg {2 * args + 4}.imm <reg> 1.inc 3.eff
0/push arg, one per arg/ <arg> <reg> 1.inc 3.eff
0/jump/ 0.reg <function>.disp 0.reg 1.inc 3.eff

If a stack reference is passed as an arg, the compiler updates the
immediate value to account for the stack changing as args are pushed.

Future work:
* There are other methods of calling that can support a wider range of
addresses. Copy jumps are limited to -64..62 bytes. This is a known
limitation that will receive more attention later.
* Address is pushed first. Not sure about this calling convention,
but it puts the vars at the top of the stack which is convenient for the
callee when using stack based args.